feat(Topic Salad): create the TopicSalad component
